Copy of skeleton from Rod
In the attempts to understand the basic works of all the abstract class systems I copied everything one by one until I get the whole schematic in my head to work with. 
(Tough as I hate to do this without previous knowledge but that' a way of learning I haven't done much. Much more practice to follow.)