Add files via upload
Working with Github and Eclipse without any instructions and old youtube tutorials is fun on its own. Working with JAVA trying to create something I have only heard of in horror stories is also fun. 

Non the less we endure and here is my half chewed penny in the meat grinder that is this lecture.

Ideology:
Make a racing game engine.

Deeds:
Swapped controls from platformer to direction based ones.
Swapped platforms to lines as I need paths not objects to stand on.
Added Directional controls.

Bugs:
Player does not rotate (haven't yet tried with "radians" coming up next)
     - Does create a stack overflow error that leaves me clueless still, but hey. At least its an error.
Player does not MOVE at all. (It didn't when I attempted any of the code to begin with so it is down to more testing.)